Refactoring. User guide written in README.md file

+++ src/main/java/com/sojerdev/forwarder/carriage/CarriageServiceImpl.java
@@ -4,12 +4,13 @@ import com.sojerdev.forwarder.carriage.driver.Driver;
 import com.sojerdev.forwarder.carriage.driver.DriverService;
 import com.sojerdev.forwarder.carriage.freight.Freight;
 import com.sojerdev.forwarder.carriage.freight.FreightService;
+import com.sojerdev.forwarder.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class CarriageServiceImpl implements CarriageService{
@@ -51,6 +52,9 @@ public class CarriageServiceImpl implements CarriageService{
 
     @Override
     public void deleteById(int id) {
+        if (findById(id) == null) {
+            throw new NotFoundException("Carriage with id - " + id + " was not found.");
+        }
         carriageRepository.deleteById(id);
     }
 
@@ -83,5 +87,26 @@ public class CarriageServiceImpl implements CarriageService{
             throw new RuntimeException("There is no driver assigned to carriage with id " + carriageId);
         }
         return driver;
+    }
+
+    @Override
+    public ResponseEntity<Object> getValues(int id) {
+        if (findById(id) == null) {
+            throw new NotFoundException("Carriage with id " + id + " was not found.");
+        }
+        List<Freight> freights = findById(id).getFreights();
+        int totalValue = 0;
+        int totalDistance = 0;
+
+        for (Freight freight: freights) {
+            totalValue += freight.getValue();
+            totalDistance += freight.getDistance();
+        }
+        Map<String, Integer> response = new HashMap<>();
+        response.put("freights",freights.size());
+        response.put("totalValue", totalValue);
+        response.put("totalDistance",totalDistance);
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 }
